Implement functions to save auto reveal service preference in localstorage
