+++ src/components/Disputes/panels/CommitPanel.js
@@ -15,6 +15,7 @@ import {
 import useOneTimeCode from '../../../hooks/useOneTimeCode'
 import { useWallet } from '../../../providers/Wallet'
 import IconOneTimeCode from '../../../assets/IconOneTimeCode.svg'
+import { saveAutoRevealPreference } from '../../../utils/crvoting-utils'
 
 const CommitPanel = React.memo(function CommitPanel({
   dispute,
@@ -32,6 +33,7 @@ const CommitPanel = React.memo(function CommitPanel({
   const handleCommit = useCallback(
     event => {
       event.preventDefault()
+      saveAutoRevealPreference(connectedAccount, revealService)
 
       onDone()
       return onCommit(
@@ -281,4 +283,5 @@ const InfoSection = React.memo(function InfoSection({ revealService }) {
     </Info>
   )
 })
+
 export default CommitPanel
